Purge the side form, and all it has touched
