+++ test/FeaturesSpec.scala
@@ -1,4 +1,4 @@
-import model.{FeatureSwitch, FeatureSwitches, InlineForm, ObscureFeed }
+import model.{FeatureSwitch, FeatureSwitches, ObscureFeed }
 import org.scalatest.{Matchers, WordSpec}
 
 class FeaturesSpec extends WordSpec with Matchers {
@@ -14,7 +14,7 @@ class FeaturesSpec extends WordSpec with Matchers {
 
       "return an updated list of feature switches" in {
         val unchangedSwitch = FeatureSwitch("test","test", enabled = true)
-        val switchToChange = InlineForm
+        val switchToChange = ObscureFeed
         val userDataSwitches: Option[List[FeatureSwitch]] = Some(List(switchToChange, unchangedSwitch))
         val remainingSwitches = FeatureSwitches.all diff List(switchToChange)
         FeatureSwitches.updateFeatureSwitchesForUser(userDataSwitches, switchToChange) should contain theSameElementsAs List(switchToChange, unchangedSwitch) ++ remainingSwitches
